backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.0
google-generativeai>=0.8.0
httpx>=0.26.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
